Adding exchange substring
strown::exchange removes a substring and replaces it with another. This
can be specified either by a substring or a character range.
strref::between returns a substring that is found between two
characters, for example " and " or < and >,
prehash.cpp: an example that loads in a source file with
PHASH("keyword") tokens that gets replaced with PHASH("keyword", <fnv1a
hash of "keyword">)